FCRE/SIRUS/Bert/requirements.txt
torch==2.6.0+cu124
transformer==4.52.4
scikit-learn==1.2.2
flash-attn --no-build-isolation
openai==1.91.0
nltk==3.9.1
retry==0.9.2